requirements.txt
numpy==2.4.4
pandas==3.0.2
scikit-learn==1.8.0
gymnasium==1.2.3
stable-baselines3==2.8.0
fastapi==0.135.2
uvicorn==0.42.0
pydantic==2.12.5
openai==2.30.0
python-dotenv==1.2.2
pyyaml==6.0.3
requests==2.33.1
slowapi==0.1.9
